refactor(server): route handlers through the Embedder + Extractor services (ENG-1747 phase 2.3)
Swaps the inline LLM/embed call sites in routes.rs to the service layer
added in 2.1-2.2, and removes the now-dead inline copies. Byte-identical
behaviour — same model, same request shapes, same parse/truncate logic.

Call-site swaps:
- analyze: extract_facts_llm(...) -> state.extractor.extract(&body.text)
  (returns ExtractedFacts { facts, raw_count } — analyze's accepted-vs-cap
  log + MAX_ANALYZE_FACTS truncation preserved).
- analyze (both the benchmark-mode sync branch and the production embed-prep
  fan-out), remember / remember_bulk prep (spawn_prepare_*_remember_job),
  ask, restore: generate_embedding(...) -> state.embedder.embed(...).
  restore's embed closure now captures Arc<dyn Embedder> instead of
  http_client + config.
- generate_recall_embedding_cached: falls through to state.embedder.embed
  on cache miss / disabled; recall_embedding_cache_key uses
  crate::services::embedder::EMBEDDING_MODEL for the cache key.
- summarize_with_prompt + ask: ChatCompletionRequest / ChatMessage /
  ChatCompletionResponse now imported from crate::services::llm_chat
  (the summarize_* path itself stays inline — it's tangled with the
  embed-prep fan-out; shape it later).

Removed from routes.rs (moved to crate::services in 2.1):
- generate_embedding fn + EmbeddingApiRequest/Response/Data structs
- EMBEDDING_MODEL const
- extract_facts_llm fn + parse_extracted_facts fn + ExtractedFacts struct
- ChatCompletion* type defs + FACT_EXTRACTION_PROMPT const
- MAX_ANALYZE_FACTS + ANALYZE_MAX_OUTPUT_TOKENS consts
- the 8 parse_extracted_facts_* tests (7 now live in extractor.rs's test
  module — the 5 originals + 2 more added here to keep full coverage)

routes.rs: 3213 -> 2846 lines (-367 net vs origin/dev). cargo check +
cargo test (135 passed; 7 parse_extracted_facts_* tests run in
services::extractor) clean — zero warnings.

Refs ENG-1747
